Rename: 도시 시니어 파일 위치 수정
